add and remove cart functionality

+++ server/db/models/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-return-await */
 const User = require('./user')
 const Wine = require('./wine')
 const Order = require('./order')
@@ -20,22 +21,48 @@ const Beer = require('./beer')
 
 Order.belongsTo(User)
 Order.hasMany(Product)
+Product.belongsTo(Order)
 
 addToCart = async (prodId, userId) => {
   const {id, name, maker, price, image} = await Product.findByPk(prodId)
-  return await Order.create({
-    userId,
-    productId: id,
-    name,
-    maker,
-    image,
-    price,
-    status: 'cart',
-    quantity: 1
-  })
+  const order = await Order.findOne({where: {userId, productId: id}})
+  if (order) {
+    return await Order.update(
+      {
+        quantity: order.quantity + 1
+      },
+      {
+        returning: true,
+        where: {
+          userId,
+          productId: id
+        }
+      }
+    )
+  } else {
+    return await Order.create({
+      userId,
+      productId: id,
+      name,
+      maker,
+      image,
+      price,
+      status: 'cart',
+      quantity: 1
+    })
+  }
 }
 removeFromCart = async (productId, userId) => {
-  return await Order.destroy({where: {userId, productId}})
+  console.log(productId, userId)
+  const order = await Order.findOne({where: {userId, productId}})
+  if (order.quantity === 1) {
+    return await Order.destroy({where: {userId, productId}})
+  } else {
+    return await Order.update(
+      {quantity: order.quantity - 1},
+      {returning: true, where: {userId, productId}}
+    )
+  }
 }
 
 module.exports = {
